+++ app/src/main/java/org/privacymatters/safespace/AboutActivity.kt
@@ -13,7 +13,7 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.updateLayoutParams
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import org.privacymatters.safespace.depracated.lib.Reload
+import org.privacymatters.safespace.utils.Reload
 
 class AboutActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
