--- app/src/main/java/org/privacymatters/safespace/depracated/lib/fileManager/FileItem.kt
@@ -1,22 +1,0 @@
-package org.privacymatters.safespace.depracated.lib.fileManager
-
-class FileItem(
-    name: String,
-    size: Long,
-    isDir: Boolean,
-    lastModified: Long
-) {
-
-    var name: String
-    var size: Long
-    var isDir: Boolean
-    var lastModified: Long
-
-    init {
-        this.name = name
-        this.size = size
-        this.isDir = isDir
-        this.lastModified = lastModified
-    }
-
-}
