--- app/src/main/java/org/privacymatters/safespace/depracated/lib/fileManager/FolderItem.kt
@@ -1,14 +1,0 @@
-package org.privacymatters.safespace.depracated.lib.fileManager
-
-class FolderItem(
-    name: String,
-    itemCount: Int,
-) {
-    var name: String
-    var itemCount: Int
-
-    init {
-        this.name = name
-        this.itemCount = itemCount
-    }
-}
