--- app/src/main/java/org/privacymatters/safespace/depracated/lib/fileManager/Operations.kt
@@ -1,679 +1,0 @@
-package org.privacymatters.safespace.depracated.lib.fileManager
-//
-//import android.app.Application
-//import android.content.Context
-//import android.content.Intent
-//import android.net.Uri
-//import android.os.FileUtils
-//import android.provider.OpenableColumns
-//import androidx.core.content.FileProvider
-//import androidx.documentfile.provider.DocumentFile
-//import kotlinx.coroutines.Dispatchers
-//import kotlinx.coroutines.withContext
-//import org.privacymatters.safespace.R
-//import org.privacymatters.safespace.experimental.main.FileHelper
-//import org.privacymatters.safespace.experimental.main.FileTransferNotification
-//import org.privacymatters.safespace.utils.Constants
-//import java.io.BufferedOutputStream
-//import java.io.File
-//import java.io.FileInputStream
-//import java.io.FileOutputStream
-//import java.io.IOException
-//import java.io.InputStream
-//import java.nio.file.Files
-//import java.nio.file.Paths
-//import java.nio.file.StandardCopyOption
-//import java.text.SimpleDateFormat
-//import java.util.Locale
-//import java.util.zip.ZipEntry
-//import java.util.zip.ZipInputStream
-//import java.util.zip.ZipOutputStream
-//
-//
-//@Suppress("unused")
-//class Operations(private val application: Application) {
-//
-//    companion object {
-//        private var internalPath: ArrayList<String> = ArrayList()
-//    }
-//
-//    private var filesList: ArrayList<FileItem> = ArrayList()
-//    private var folderList: ArrayList<FolderItem> = ArrayList()
-//
-//    var moveFileFrom: String? = null
-//    var moveFileTo: String? = null
-//
-//    fun getFilesDir(): String {
-//        // root folder inside app files directory will be the first folder
-//        return application.filesDir.canonicalPath.toString() + File.separator + Constants.ROOT
-//    }
-//
-//    fun getInternalPath(): String {
-//        return internalPath.joinToString(File.separator)
-//    }
-//
-//    fun setInternalPath(dir: String) {
-//        if (internalPath.isEmpty() || internalPath.last() != dir) {
-//            internalPath.add(dir)
-//        }
-//    }
-//
-//    fun setGetPreviousAndCurrentPath(): Pair<String, String> {
-//        val lastPath = internalPath.last()
-//        var currentPath = ""
-//
-//        if (internalPath.isNotEmpty()) {
-//            internalPath.removeLast()
-//            currentPath = if (internalPath.size > 0) internalPath.last() else ""
-//        }
-//        return Pair(lastPath, currentPath)
-//    }
-//
-//    fun isRootDirectory(): Boolean {
-//        return internalPath.isEmpty()
-//    }
-//
-//    fun joinPath(vararg pathList: String): String {
-//
-//        return pathList.joinToString(File.separator).replace("//", "/")
-//
-//    }
-//
-//    private suspend fun copyFileWithProgressNotifications(
-//        sourceFileName: String,
-//        sourceFileStream: InputStream,
-//        targetFileStream: FileOutputStream,
-//        type: FileTransferNotification.NotificationType,
-//        context: Context
-//    ) {
-//        val uniqueNotificationId = sourceFileName.hashCode()
-//        val fileTransferNotification = FileTransferNotification(context, uniqueNotificationId)
-//        try {
-//            FileHelper.copyFileWithProgress(sourceFileStream, targetFileStream)
-//                .collect { progress ->
-//                    fileTransferNotification.showProgressNotification(
-//                        fileName = sourceFileName,
-//                        progress = progress,
-//                        type = type
-//                    )
-//                }
-//            fileTransferNotification.showSuccessNotification(
-//                fileName = sourceFileName,
-//                type = type
-//            )
-//        } catch (exception: Exception) {
-//            exception.printStackTrace()
-//            fileTransferNotification.showFailureNotification(sourceFileName, exception)
-//        }
-//    }
-//
-//    suspend fun importFile(uri: Uri, internalPath: String): Int =
-//        withContext(Dispatchers.IO) {
-//            try {
-//                // create directory if not exists
-//                // createDir(internalPath, "")
-//                var sourceFileName = ""
-//                val cursor = application.contentResolver.query(
-//                    uri, null, null, null, null, null
-//                )
-//
-//                cursor?.use {
-//                    if (it.moveToFirst()) {
-//                        val colIndex = it.getColumnIndex(OpenableColumns.DISPLAY_NAME)
-//                        if (colIndex >= 0) {
-//                            sourceFileName = it.getString(colIndex)
-//                        }
-//                    }
-//                }
-//
-//                // byte array of source file
-//                val sourceFileStream = application.contentResolver.openInputStream(uri)
-//
-//                // output stream for target file
-//                val targetFileStream =
-//                    FileOutputStream(
-//                        File(joinPath(getFilesDir(), internalPath, sourceFileName))
-//                    )
-//
-//                if (sourceFileStream != null) {
-//                    copyFileWithProgressNotifications(
-//                        sourceFileName, sourceFileStream,
-//                        targetFileStream,
-//                        FileTransferNotification.NotificationType.Import,
-//                        application.applicationContext,
-//                    )
-//                } else {
-//                    return@withContext -1
-//                }
-//
-//                sourceFileStream.close()
-//                targetFileStream.close()
-//
-//            } catch (e: Exception) {
-//                return@withContext -1
-//            }
-//
-//            return@withContext 1
-//        }
-//
-//    fun createDir(internalPath: String, newDirName: String): Int {
-//
-//        try {
-//            val dirPath = joinPath(getFilesDir(), internalPath, newDirName)
-//
-//            val newDir = File(dirPath)
-//
-//            if (!newDir.exists()) {
-//                newDir.mkdirs()
-//            }
-//
-//        } catch (e: FileSystemException) {
-//            return 0
-//        }
-//
-//        return 1
-//
-//    }
-//
-//    fun initRootDir(): Int {
-//
-//        try {
-//
-//            val newDir = File(getFilesDir())
-//
-//            if (!newDir.exists()) {
-//                newDir.mkdirs()
-//            }
-//
-//        } catch (e: FileSystemException) {
-//            return 0
-//        }
-//
-//        return 1
-//
-//    }
-//
-//    fun getContents(internalPath: String): Pair<List<FileItem>, List<FolderItem>> {
-//
-//        val dirPath = File(joinPath(getFilesDir(), internalPath))
-//
-//        val contents = dirPath.listFiles()
-//
-//        filesList.clear()
-//        folderList.clear()
-//
-//        for (item in contents!!) {
-//            if (item.isDirectory) {
-//
-//                val fileInsideFolder = item.listFiles()
-//
-//                val fileCount = fileInsideFolder?.size ?: 0
-//
-//                folderList.add(FolderItem(item.name, fileCount))
-//
-//            } else {
-//
-//                filesList.add(
-//                    FileItem(
-//                        item.name,
-//                        item.length(),
-//                        item.isDirectory,
-//                        item.lastModified()
-//                    )
-//                )
-//            }
-//        }
-//
-//        // sort -> folders first -> ascending by name
-//        filesList.sortWith(compareByDescending<FileItem> { it.isDir }.thenBy { it.name })
-//
-//        return Pair(filesList, folderList)
-//
-//    }
-//
-//    fun renameFile(file: FileItem, internalPath: String, newFileName: String): Int {
-//
-//        try {
-//            val absolutePath = joinPath(getFilesDir(), internalPath, File.separator)
-//
-//            val absoluteFilePathOld = File(absolutePath + file.name)
-//
-//            val absoluteFilePathNew = File(absolutePath + newFileName)
-//
-//            absoluteFilePathOld.renameTo(absoluteFilePathNew)
-//
-//        } catch (e: Exception) {
-//            return 0
-//        }
-//
-//        return 1
-//    }
-//
-//    fun deleteFile(file: FileItem, internalPath: String): Int {
-//        try {
-//            val fileToDelete = File(joinPath(getFilesDir(), internalPath, file.name))
-//
-//            if (fileToDelete.exists()) {
-//                if (file.isDir) {
-//                    deleteDirectory(fileToDelete)
-//                }
-//                fileToDelete.delete()
-//            }
-//
-//        } catch (e: Exception) {
-//            return 0
-//        }
-//
-//        return 1
-//
-//    }
-//
-//    fun deleteFolder(folder: FolderItem, internalPath: String): Int {
-//        try {
-//            val folderToDelete = File(joinPath(getFilesDir(), internalPath, folder.name))
-//
-//            if (folderToDelete.exists()) {
-//                deleteDirectory(folderToDelete)
-//                folderToDelete.delete()
-//            }
-//
-//        } catch (e: Exception) {
-//            return 0
-//        }
-//
-//        return 1
-//
-//    }
-//
-//    private fun deleteDirectory(fileToDelete: File): Int {
-//        try {
-//            val dirContents = fileToDelete.listFiles()
-//            for (file in dirContents!!) {
-//                if (file.isDirectory) {
-//                    deleteDirectory(File(file.absolutePath))
-//                } else {
-//                    file.delete()
-//                }
-//            }
-//            fileToDelete.delete()
-//
-//        } catch (e: Exception) {
-//            return -1
-//        }
-//        return 1
-//    }
-//
-//    fun moveFile(): Int {
-//
-//        try {
-//
-//            Files.move(
-//                Paths.get(moveFileFrom),
-//                Paths.get(moveFileTo),
-//                StandardCopyOption.REPLACE_EXISTING
-//            )
-//
-//        } catch (e: Exception) {
-//            return -1
-//        } finally {
-//
-//            moveFileFrom = null
-//            moveFileTo = null
-//        }
-//        return 1
-//    }
-//
-//    fun copyFile(): Int {
-//
-//        try {
-//
-//            // byte array of source file
-//            val sourceFileStream = FileInputStream(moveFileFrom)
-//
-//            // output stream for target file
-//            val targetFileStream = FileOutputStream(moveFileTo)
-//
-//            FileUtils.copy(sourceFileStream, targetFileStream)
-//
-//            sourceFileStream.close()
-//            targetFileStream.close()
-//
-//        } catch (e: Exception) {
-//            return -1
-//        } finally {
-//
-//            moveFileFrom = null
-//            moveFileTo = null
-//        }
-//        return 1
-//    }
-//
-//    fun shareFile(file: FileItem, internalPath: String) {
-//        val fileToShare = File(joinPath(getFilesDir(), internalPath, file.name))
-//        val fileUri = FileProvider.getUriForFile(
-//            application,
-//            application.applicationContext.packageName + ".provider",
-//            fileToShare
-//        )
-//
-//        val intent = Intent(Intent.ACTION_SEND)
-//        intent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-//        intent.type = application.contentResolver.getType(fileUri)
-//        intent.putExtra(Intent.EXTRA_STREAM, fileUri)
-//
-//        val chooser = Intent.createChooser(
-//            intent,
-//            application.getString(R.string.share_chooser_title)
-//        )
-//        chooser.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
-//        application.startActivity(chooser)
-//    }
-//
-//    fun createTextNote(noteName: String): String {
-//
-//        val filePath = joinPath(getFilesDir(), getInternalPath(), noteName)
-//
-//        val newTextNote = File(filePath)
-//        val result = newTextNote.createNewFile()
-//
-//        return if (result) {
-//            filePath
-//        } else {
-//            Constants.FILE_EXIST
-//        }
-//
-//    }
-//
-//    suspend fun exportItems(
-//        exportUri: Uri,
-//        selectedItem: FileItem
-//    ): Boolean = withContext(Dispatchers.IO) {
-//
-//        try {
-//            val fileToExport = File(getFilesDir() + File.separator + getInternalPath() + File.separator + selectedItem.name)
-//
-//            val fis = FileInputStream(fileToExport)
-//
-//            val directory = DocumentFile.fromTreeUri(application, exportUri)
-//            val file = directory!!.createFile("*", selectedItem.name)
-//            val pfd = application.contentResolver.openFileDescriptor(file!!.uri, "w")
-//            val fos = FileOutputStream(pfd!!.fileDescriptor)
-//
-//            copyFileWithProgressNotifications(
-//                selectedItem.name, fis, fos,
-//                FileTransferNotification.NotificationType.Export,
-//                application.applicationContext
-//            )
-//
-//            fos.close()
-//            pfd.close()
-//
-//
-//        } catch (e: Exception) {
-//            return@withContext false
-//        }
-//        return@withContext true
-//    }
-//
-//    fun isPreviousRootDirectory(): Boolean {
-//        return internalPath.size == 0
-//    }
-//
-//    fun exportBackup(exportUri: Uri): Int {
-//
-//        try {
-//            val backupName =
-//                "SafeSpace-" + SimpleDateFormat("yyyy-MM-dd-HH-mm-ss", Locale.getDefault())
-//                    .format(System.currentTimeMillis()) + ".zip"
-//
-//            val directory = DocumentFile.fromTreeUri(application, exportUri)
-//            val file = directory!!.createFile("application/zip", backupName)
-//            val pfd = application.contentResolver.openFileDescriptor(file!!.uri, "w")
-//
-//            val inputDirectory = File(getFilesDir())
-//
-//            ZipOutputStream(BufferedOutputStream(FileOutputStream(pfd!!.fileDescriptor))).use { zos ->
-//
-//                inputDirectory.walkTopDown().forEach { file ->
-//                    val zipFileName = Constants.ROOT +
-//                            file.absolutePath.removePrefix(inputDirectory.absolutePath)
-//
-//                    val entry = ZipEntry("$zipFileName${(if (file.isDirectory) "/" else "")}")
-//
-//                    zos.putNextEntry(entry)
-//
-//                    if (file.isFile) {
-//                        file.inputStream().use { fis -> fis.copyTo(zos) }
-//                    }
-//                }
-//
-//            }
-//            pfd.close()
-//
-//        } catch (e: IOException) {
-////            e.printStackTrace()
-//            return if (e.message.toString().lowercase().contains("no space left")) {
-//                4
-//            } else {
-//                1
-//            }
-//        }
-//        return 0
-//    }
-//
-//    @Throws(SecurityException::class)
-//    fun importBackup(backupUri: Uri): Int {
-//
-//        try {
-//
-//            // byte array of source file
-//            val sourceFileStream = application.contentResolver.openInputStream(backupUri)
-//
-//            val zis = ZipInputStream(sourceFileStream)
-//
-//            var zipEntry = zis.nextEntry
-//
-//            while (zipEntry != null) {
-//                val newFile = File(getFilesDir(), zipEntry.name)
-//
-//                // https://support.google.com/faqs/answer/9294009
-//                val canonicalPath = newFile.canonicalPath
-//
-//                if (!canonicalPath.startsWith(getFilesDir())) {
-//                    throw SecurityException()
-//                } else {
-//                    if (zipEntry.isDirectory) {
-//                        if (!newFile.isDirectory && !newFile.mkdirs()) {
-//                            throw IOException("Failed to create directory $newFile")
-//                        }
-//                    } else {
-//                        // fix for Windows-created archives
-//                        val parent = newFile.parentFile
-//                        if (parent != null) {
-//                            if (!parent.isDirectory && !parent.mkdirs()) {
-//                                throw IOException("Failed to create directory $parent")
-//                            }
-//                        }
-//
-//                        // write file content
-//                        val fos = FileOutputStream(newFile)
-//
-//                        zis.copyTo(fos)
-//
-//                        fos.close()
-//                    }
-//                }
-//                zipEntry = zis.nextEntry
-//            }
-//
-//            zis.closeEntry()
-//            zis.close()
-//            sourceFileStream?.close()
-//
-//
-//        } catch (exc: IOException) {
-//            return if (exc.message.toString().lowercase().contains("no space left")) {
-//                4
-//            } else {
-//                1
-//            }
-//        } catch (sec: SecurityException) {
-////            sec.printStackTrace()
-//            return 1
-//        }
-//
-//        try {
-//
-//            // android 14 throws security exception if zip archive has files at / location
-//            // therefore after extracting the zip files all files will be moved out of the first folder in the zip
-//
-//            val importDir = File(joinPath(getFilesDir(), Constants.ROOT))
-//
-//            importDir.walkTopDown().forEach { file ->
-//
-//                val sourcePath = file.absolutePath
-//                val targetPath = file.absolutePath.replaceFirst("/root", "")
-//
-//                if (file.isDirectory) {
-//                    File(targetPath).mkdirs()
-//                } else {
-//                    Files.move(
-//                        Paths.get(sourcePath),
-//                        Paths.get(targetPath),
-//                        StandardCopyOption.REPLACE_EXISTING
-//                    )
-//                }
-//            }
-//        } catch (exc: Exception) {
-//            exc.printStackTrace()
-//            return 1
-//        } finally {
-//            File(joinPath(getFilesDir(), Constants.ROOT)).deleteRecursively()
-//        }
-//
-//        return 0
-//    }
-//
-//
-//    @Suppress("Unused")
-//    private fun recursiveDirectoryRead(
-//        path: String,
-//        pFilesArray: ArrayList<String>
-//    ): ArrayList<String> {
-//
-//        var filesArray = pFilesArray
-//
-//        val directoryContents = File(path).listFiles()
-//
-//        for (item in directoryContents!!) {
-//            val filePath = joinPath(path, item.name)
-//            if (item.isDirectory) {
-//                filesArray = recursiveDirectoryRead(filePath, filesArray)
-//            } else {
-//                filesArray.add(filePath)
-//            }
-//        }
-//
-//        return filesArray
-//
-//    }
-//
-//    fun compressFolder(folder: FolderItem): Int {
-//        try {
-//            val zipName = joinPath(getFilesDir(), getInternalPath(), folder.name + ".zip")
-//
-//            val inputDirectory = File(joinPath(getFilesDir(), getInternalPath(), folder.name))
-//
-//            ZipOutputStream(BufferedOutputStream(FileOutputStream(zipName))).use { zos ->
-//
-//                inputDirectory.walkTopDown().forEach { file ->
-//                    val zipFileName = folder.name +
-//                            file.absolutePath.removePrefix(inputDirectory.absolutePath)
-//
-//                    val entry = ZipEntry("$zipFileName${(if (file.isDirectory) "/" else "")}")
-//
-//                    zos.putNextEntry(entry)
-//
-//                    if (file.isFile) {
-//                        file.inputStream().use { fis -> fis.copyTo(zos) }
-//                    }
-//                }
-//
-//            }
-//
-//        } catch (e: IOException) {
-////            e.printStackTrace()
-//            return if (e.message.toString().lowercase().contains("no space left")) {
-//                4
-//            } else {
-//                1
-//            }
-//        }
-//        return 0
-//    }
-//
-//    @Throws(SecurityException::class)
-//    fun extractZip(filePath: String): Int {
-//        try {
-//
-//            val currentDir = joinPath(getFilesDir(), getInternalPath())
-//
-//            // byte array of source file
-//            val sourceFileStream = FileInputStream(filePath)
-//
-//            val zis = ZipInputStream(sourceFileStream)
-//
-//            var zipEntry = zis.nextEntry
-//
-//            while (zipEntry != null) {
-//                val newFile = File(currentDir, zipEntry.name)
-//
-//                // https://support.google.com/faqs/answer/9294009
-//                val canonicalPath = newFile.canonicalPath
-//
-//                if (!canonicalPath.startsWith(getFilesDir())) {
-//                    throw SecurityException()
-//                } else {
-//                    if (zipEntry.isDirectory) {
-//                        if (!newFile.isDirectory && !newFile.mkdirs()) {
-//                            throw IOException("Failed to create directory $newFile")
-//                        }
-//                    } else {
-//                        // fix for Windows-created archives
-//                        val parent = newFile.parentFile
-//                        if (parent != null) {
-//                            if (!parent.isDirectory && !parent.mkdirs()) {
-//                                throw IOException("Failed to create directory $parent")
-//                            }
-//                        }
-//
-//                        // write file content
-//                        val fos = FileOutputStream(newFile)
-//
-//                        zis.copyTo(fos)
-//
-//                        fos.close()
-//                    }
-//                }
-//                zipEntry = zis.nextEntry
-//            }
-//
-//            zis.closeEntry()
-//            zis.close()
-//            sourceFileStream.close()
-//        } catch (exc: IOException) {
-//            return if (exc.message.toString().lowercase().contains("no space left")) {
-//                4
-//            } else {
-//                1
-//            }
-//        } catch (sec: SecurityException) {
-////            sec.printStackTrace()
-//            return 1
-//
-//
-//        }
-//        return 0
-//    }
-//
-//}
