+++ app/src/main/java/org/privacymatters/safespace/document/PDFView.kt
@@ -14,8 +14,8 @@ import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.SnapHelper
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import org.privacymatters.safespace.R
-import org.privacymatters.safespace.experimental.main.DataManager
-import org.privacymatters.safespace.depracated.lib.Reload
+import org.privacymatters.safespace.main.DataManager
+import org.privacymatters.safespace.utils.Reload
 import org.privacymatters.safespace.utils.Constants
 import org.privacymatters.safespace.utils.Utils
 import java.io.File
