--- app/src/main/java/org/privacymatters/safespace/experimental/main/Item.kt
@@ -1,13 +1,0 @@
-package org.privacymatters.safespace.experimental.main
-
-import java.util.UUID
-
-data class Item(
-    val id : UUID,
-    val name: String,
-    val size: Long,
-    val isDir: Boolean,
-    val itemCount: String,
-    val lastModified: Long,
-    val isSelected: Boolean,
-)
