--- app/src/main/java/org/privacymatters/safespace/experimental/main/ui/Color.kt
@@ -1,74 +1,0 @@
-package org.privacymatters.safespace.experimental.main.ui
-import androidx.compose.ui.graphics.Color
-
-val primaryLight = Color(0xFF35618E)
-val onPrimaryLight = Color(0xFFFFFFFF)
-val primaryContainerLight = Color(0xFFD1E4FF)
-val onPrimaryContainerLight = Color(0xFF001D36)
-val secondaryLight = Color(0xFF535F70)
-val onSecondaryLight = Color(0xFFFFFFFF)
-val secondaryContainerLight = Color(0xFFD6E4F7)
-val onSecondaryContainerLight = Color(0xFF0F1C2B)
-val tertiaryLight = Color(0xFF6B5778)
-val onTertiaryLight = Color(0xFFFFFFFF)
-val tertiaryContainerLight = Color(0xFFF2DAFF)
-val onTertiaryContainerLight = Color(0xFF251432)
-val errorLight = Color(0xFFBA1A1A)
-val onErrorLight = Color(0xFFFFFFFF)
-val errorContainerLight = Color(0xFFFFDAD6)
-val onErrorContainerLight = Color(0xFF410002)
-val backgroundLight = Color(0xFFF8F9FF)
-val onBackgroundLight = Color(0xFF191C20)
-val surfaceLight = Color(0xFFF8F9FF)
-val onSurfaceLight = Color(0xFF191C20)
-val surfaceVariantLight = Color(0xFFDFE2EB)
-val onSurfaceVariantLight = Color(0xFF42474E)
-val outlineLight = Color(0xFF73777F)
-val outlineVariantLight = Color(0xFFC3C7CF)
-val scrimLight = Color(0xFF000000)
-val inverseSurfaceLight = Color(0xFF2E3135)
-val inverseOnSurfaceLight = Color(0xFFEFF0F7)
-val inversePrimaryLight = Color(0xFFA0CAFD)
-val surfaceDimLight = Color(0xFFD8DAE0)
-val surfaceBrightLight = Color(0xFFF8F9FF)
-val surfaceContainerLowestLight = Color(0xFFFFFFFF)
-val surfaceContainerLowLight = Color(0xFFF2F3F9)
-val surfaceContainerLight = Color(0xFFECEEF4)
-val surfaceContainerHighLight = Color(0xFFE6E8EE)
-val surfaceContainerHighestLight = Color(0xFFE1E2E8)
-
-val primaryDark = Color(0xFFA0CAFD)
-val onPrimaryDark = Color(0xFF003258)
-val primaryContainerDark = Color(0xFF184975)
-val onPrimaryContainerDark = Color(0xFFD1E4FF)
-val secondaryDark = Color(0xFFBAC8DB)
-val onSecondaryDark = Color(0xFF253140)
-val secondaryContainerDark = Color(0xFF3B4858)
-val onSecondaryContainerDark = Color(0xFFD6E4F7)
-val tertiaryDark = Color(0xFFD6BEE5)
-val onTertiaryDark = Color(0xFF3B2948)
-val tertiaryContainerDark = Color(0xFF523F5F)
-val onTertiaryContainerDark = Color(0xFFF2DAFF)
-val errorDark = Color(0xFFFFB4AB)
-val onErrorDark = Color(0xFF690005)
-val errorContainerDark = Color(0xFF93000A)
-val onErrorContainerDark = Color(0xFFFFDAD6)
-val backgroundDark = Color(0xFF000000)
-val onBackgroundDark = Color(0xFFE1E2E8)
-val surfaceDark = Color(0xFF101418)
-val onSurfaceDark = Color(0xFFE1E2E8)
-val surfaceVariantDark = Color(0xFF42474E)
-val onSurfaceVariantDark = Color(0xFFC3C7CF)
-val outlineDark = Color(0xFF8D9199)
-val outlineVariantDark = Color(0xFF42474E)
-val scrimDark = Color(0xFF000000)
-val inverseSurfaceDark = Color(0xFFE1E2E8)
-val inverseOnSurfaceDark = Color(0xFF2E3135)
-val inversePrimaryDark = Color(0xFF35618E)
-val surfaceDimDark = Color(0xFF101418)
-val surfaceBrightDark = Color(0xFF36393E)
-val surfaceContainerLowestDark = Color(0xFF0B0E13)
-val surfaceContainerLowDark = Color(0xFF191C20)
-val surfaceContainerDark = Color(0xFF1D2024)
-val surfaceContainerHighDark = Color(0xFF272A2F)
-val surfaceContainerHighestDark = Color(0xFF32353A)
