+++ app/src/main/java/org/privacymatters/safespace/main/MainnActivity.kt
@@ -1,4 +1,4 @@
-package org.privacymatters.safespace.experimental.main
+package org.privacymatters.safespace.main
 
 import android.Manifest
 import android.content.Intent
@@ -38,11 +38,11 @@ import androidx.core.content.ContextCompat
 import androidx.lifecycle.lifecycleScope
 import kotlinx.coroutines.launch
 import org.privacymatters.safespace.R
-import org.privacymatters.safespace.depracated.lib.Reload
-import org.privacymatters.safespace.experimental.main.ui.BottomAppBar
-import org.privacymatters.safespace.experimental.main.ui.ItemList
-import org.privacymatters.safespace.experimental.main.ui.SafeSpaceTheme
-import org.privacymatters.safespace.experimental.main.ui.TopAppBar
+import org.privacymatters.safespace.utils.Reload
+import org.privacymatters.safespace.main.ui.BottomAppBar
+import org.privacymatters.safespace.main.ui.ItemList
+import org.privacymatters.safespace.main.ui.SafeSpaceTheme
+import org.privacymatters.safespace.main.ui.TopAppBar
 
 class MainnActivity : AppCompatActivity() {
