+++ app/src/main/java/org/privacymatters/safespace/media/MediaActivity.kt
@@ -12,7 +12,7 @@ import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import androidx.viewpager2.widget.ViewPager2
 import org.privacymatters.safespace.R
-import org.privacymatters.safespace.depracated.lib.Reload
+import org.privacymatters.safespace.utils.Reload
 
 class MediaActivity : AppCompatActivity() {
